Refactor: Split Rust node API into smaller submodules
