+++ apis/rust/node/Cargo.toml
@@ -23,6 +23,7 @@ flume = "0.10.14"
 uuid = { version = "1.1.2", features = ["v4"] }
 capnp = "0.14.11"
 bincode = "1.3.3"
+shared_memory = "0.12.0"
 
 [dev-dependencies]
 tokio = { version = "1.24.2", features = ["rt"] }
