+++ apis/rust/node/src/lib.rs
@@ -1,8 +1,10 @@
-pub use daemon_connection::{Event, EventStream};
+pub use daemon_connection::EventStream;
 pub use dora_core;
 pub use dora_core::message::{uhlc, Metadata, MetadataParameters};
+pub use event::{Event, MappedInputData};
 pub use flume::Receiver;
 pub use node::DoraNode;
 
 mod daemon_connection;
+mod event;
 mod node;
